Revert "Converted Value Arrays to Int"
This reverts commit a1130fa588ef6c86ca2a4a06bdf866be57ebd028.

+++ OpenCL Tutorials/Tutorial 3/my_kernels3.cl
@@ -1,5 +1,5 @@
 ﻿//minimum value reduce
-__kernel void value_min(__global const int* A, __global int* B, __local int* min) {
+__kernel void value_min(__global const float* A, __global float* B, __local float* min) {
 	int id = get_global_id(0);
 	int lid = get_local_id(0);
 	int N = get_local_size(0);
@@ -27,7 +27,7 @@ __kernel void value_min(__global const int* A, __global int* B, __local int* min
 }
 
 //maximum value reduce
-__kernel void value_max(__global const int* A, __global int* B, __local int* max) {
+__kernel void value_max(__global const float* A, __global float* B, __local float* max) {
 	int id = get_global_id(0);
 	int lid = get_local_id(0);
 	int N = get_local_size(0);
@@ -55,7 +55,7 @@ __kernel void value_max(__global const int* A, __global int* B, __local int* max
 }
 
 //additive value reduce for average
-__kernel void value_avg(__global const int* A, __global float* B, __local float* avg) {
+__kernel void value_avg(__global const float* A, __global float* B, __local float* avg) {
 	int id = get_global_id(0);
 	int lid = get_local_id(0);
 	int N = get_local_size(0);
